src/utils/gameLogic.ts
import { BoardState, Player, WinningCombination } from '../types/game';

const WINNING_COMBINATIONS: WinningCombination[] = [
  [0, 1, 2], [3, 4, 5], [6, 7, 8], // Rows
  [0, 3, 6], [1, 4, 7], [2, 5, 8], // Columns
  [0, 4, 8], [2, 4, 6] // Diagonals
];

export const checkWinner = (board: BoardState): Player | null => {
  for (const combination of WINNING_COMBINATIONS) {
    const [a, b, c] = combination;
    if (board[a] && board[a] === board[b] && board[a] === board[c]) {
      return board[a];
    }
  }
  return null;
};

export const checkDraw = (board: BoardState): boolean => {
  return board.every(cell => cell !== null);
};

export const getInitialBoard = (): BoardState => Array(9).fill(null);